requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Werkzeug
openai
numpy
sentence-transformers
tiktoken
python-dotenv
PyPDF2
python-docx
beautifulsoup4
requests
tqdm
youtube-transcript-api 
gunicorn 
flask-migrate
selenium
langdetect
flask-migrate

